API/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlmodel==0.0.22
passlib[argon2]==1.7.4
python-jose==3.3.0
openai==1.51.0
python-dotenv==1.0.1
sqlalchemy==2.0.38
aiomysql==0.2.0
pydantic-settings==2.6.1







